extracted out code from privateChats.ts to dedicated files
