+++ src/hooks/chats/privateChats/useOnlineUsersToDisplay.ts
@@ -1,0 +1,16 @@
+import { useRecentWorldUsers } from "../../users";
+import { useUser } from "../../useUser";
+import { useMemo } from "react";
+
+export const useOnlineUsersToDisplay = () => {
+  const { recentWorldUsers } = useRecentWorldUsers();
+  const { user } = useUser();
+
+  const userId = user?.uid;
+
+  // Filter out self
+  return useMemo(() => recentWorldUsers.filter((user) => user.id !== userId), [
+    recentWorldUsers,
+    userId,
+  ]);
+};
